DijkstraAlgorithm

+++ 06.BaekJoon/BaekJoon/Practice.cpp
@@ -1,9 +1,7 @@
 #include <iostream>
-#include <algorithm>
-#include <vector>
 using namespace std;
 
-int A[200001];
+long long int A[200001];
 
 int main(){
 	ios_base :: sync_with_stdio(false);
@@ -13,7 +11,7 @@ int main(){
 	int N, M, num = 0;
 	cin >> N >> M;
 	
-	int a, b;
+	long long int a, b;
 	for(int i = 0; i < N; i++){
 		cin >> a;
 		A[a]++;
@@ -23,6 +21,7 @@ int main(){
 		A[b]++;
 	} 	
 	
+	// O(n)
 	for(int i = 0; i < 200001; i++){
 		if(A[i] == 1){
 			num++;
